Fix default filter logic in ServeUI method
The default filter logic in the ServeUI method has been fixed to include all filters (city, country, service, and version) when checking if the "From" and "To" filters are zero. This ensures that the default filter is applied correctly when all other filters are empty.

This commit addresses the following issue:
- The default filter logic was not correctly considering all filters when determining if the "From" and "To" filters are zero.

Signed-off-by: SammyOina <sammyoina@gmail.com>